Add files via upload
The main objective of the game is for the snake to eat food, resulting in its growth.
The player controls the snake's movement and must navigate within a defined grid
space while avoiding collisions with walls, itself, or randomly generated obstacles. If
a collision occurs, the game ends. The rules of the game are straightforward and easy
to understand, allowing novice players to quickly get the hang of it. At the start of the
game, the snake's length is one segment, and with each food consumed, its length
increases, thereby raising the difficulty level.
The core of the game design is to balance challenge with enjoyment. In the initial
phases, the food generation location is relatively simple, but as the game progresses,
the number and distribution of obstacles increase, requiring players or agents to
continuously adjust their strategies to adapt to the changing environment. To enhance
the game's fun factor, we also designed multiple colors and shapes for the food and
obstacles, making the game visually appealing